+++ src/Microsoft.Restier.Security/Microsoft.Restier.Security.csproj
@@ -32,6 +32,7 @@
       <Link>Shared\Ensure.cs</Link>
     </Compile>
     <Compile Include="ApiPermission.cs" />
+    <Compile Include="ApiPolicyActivator.cs" />
     <Compile Include="AssertRoleAttribute.cs" />
     <Compile Include="ApiConfigurationExtensions.cs" />
     <Compile Include="ApiPermissionType.cs" />
